controller

+++ portfolioprojects/funiturestore/backend/controllers/product.controller.js
@@ -71,6 +71,20 @@ export async function deleteProduct(req, res) {
         error: "Product id provided does not exists in the catalogue",
       });
 
+    const product = await Product.findById(id);
+
+    if (!product) {
+      return res.status(404).json({ error: "Product not found" });
+    }
+
+    // Extract the public ID from the image URL
+    if (product.img) {
+      const oldImageUrl = product.img;
+      const oldImagePublicId = oldImageUrl.split("/").pop().split(".")[0];
+
+      await cloudinary.uploader.destroy(oldImagePublicId);
+    }
+
     await Product.findByIdAndDelete(id);
 
     res.status(200).json({ message: "Product deleted successfully" });
@@ -82,17 +96,37 @@ export async function deleteProduct(req, res) {
 
 export async function updateProduct(req, res) {
   try {
-    const { name, price, description, category, productPic, quantity } =
-      req.body;
+    const { name, price, description, category, brand, quantity } = req.body;
+    let { img } = req.body;
     const { id } = req.params;
 
     if (!name || !price || !description || !category) {
       return res.status(400).json({ error: "Enter all required fields" });
     }
 
+    const existingProduct = await Product.findById(id);
+
+    if (!existingProduct)
+      return res.status(404).json({ error: "Product not Found" });
+
+    if (!existingProduct.img && img) {
+      const uploadedResponse = await cloudinary.uploader.upload(img);
+      img = uploadedResponse.secure_url;
+    } else if (img && img !== existingProduct.img) {
+      const oldImageUrl = existingProduct.img;
+      const oldImagePublicId = oldImageUrl.split("/").pop().split(".")[0];
+
+      await cloudinary.uploader.destroy(oldImagePublicId);
+
+      const uploadedResponse = await cloudinary.uploader.upload(img);
+      img = uploadedResponse.secure_url;
+    } else {
+      img = existingProduct.img;
+    }
+
     const updatedProduct = await Product.findByIdAndUpdate(
       id,
-      { name, price, description, category, productPic, quantity },
+      { name, price, description, category, img, quantity, brand },
       { new: true, runValidators: true }
     );
